/fix shimmer loading final form
